Refactor dataset handling
