+++ data/data_goes_here.md
@@ -1,3 +1,9 @@
 # Place the downloaded data in this directory
 Download the data from the following link:
 https://www.kaggle.com/datasets/jeanmidev/smart-meters-in-london
+### Expected look of the directory:
+```
+data/
+    ├── smart-meters-in-london.zip
+    └── data_goes_here.md
+```
