+++ src/scripts/__init__.py
@@ -1,0 +1,5 @@
+from .dataset_validator import handle_dataset_availability
+
+__all__ = [
+    handle_dataset_availability,
+]
